MotorUniversalV2/backend/requirements.txt
# Core
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Database
psycopg2-binary==2.9.9
pymssql==2.2.11
alembic==1.13.0

# Redis (cache & sessions)
redis==5.0.1
Flask-Caching==2.1.0

# Security
argon2-cffi==23.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Azure
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Utils
Pillow==10.1.0
qrcode[pil]==7.4.2
python-dateutil==2.8.2
email-validator==2.1.0

# API Documentation
flasgger==0.9.7.1

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Development
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==21.0.0
